fix: repair factory_visualization test build after main merge
Replace undefined projectionStub with recordingsstub.Service and dedupe import-boundary test helpers that collided when merging main.
