+++ pkg/services/factory_visualization/runtime_import_boundary_test.go
@@ -6,11 +6,7 @@ import (
 	"testing"
 )
 
-const (
-	modulePrefix            = "github.com/portpowered/infinite-you/"
-	factoryRuntimeRoot      = modulePrefix + "pkg/services/factory_runtime"
-	factoryVisualizationRoot = modulePrefix + "pkg/services/factory_visualization"
-)
+const factoryRuntimeRoot = modulePrefix + "pkg/services/factory_runtime"
 
 // TestProductionPackagesImportFactoryRuntimeRootOnly seals CUT-VIS-RUN story 001:
 // Factory Visualization production packages may depend on Factory Runtime only
@@ -26,17 +22,6 @@ func TestProductionPackagesImportFactoryRuntimeRootOnly(t *testing.T) {
 			assertProductionImportsUseRuntimeRootOnly(t, pkg)
 		})
 	}
-}
-
-func listFactoryVisualizationPackages(t *testing.T) []string {
-	t.Helper()
-
-	cmd := exec.Command("go", "list", factoryVisualizationRoot+"/...")
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		t.Fatalf("go list factory visualization packages: %v\n%s", err, output)
-	}
-	return strings.Fields(string(output))
 }
 
 func assertProductionImportsUseRuntimeRootOnly(t *testing.T, packagePath string) {
@@ -77,15 +62,4 @@ func isForbiddenFactoryVisualizationRuntimeImport(importPath string) bool {
 		return true
 	}
 	return false
-}
-
-func shortFactoryVisualizationPackageName(packagePath string) string {
-	if strings.HasPrefix(packagePath, factoryVisualizationRoot) {
-		rest := strings.TrimPrefix(packagePath, factoryVisualizationRoot)
-		if rest == "" {
-			return "factory_visualization"
-		}
-		return strings.TrimPrefix(rest, "/")
-	}
-	return packagePath
 }
